# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: Invalid credentials or user not in allowed group
        type: string
    type: object
...
  models.LogsResponse:
    properties:
      containerName:
        type: string
      logs:
        type: string
    type: object
...
  models.TopologyDocEventResponse:
    properties:
      action:
        example: change
        type: string
      documentKind:
        example: annotations
        type: string
      labName:
        example: mylab
        type: string
      path:
        example: mylab.clab.yml.annotations.json
        type: string
      revision:
        example: yaml=123-456;annotations=78-910
        type: string
      type:
        example: topology-doc
        type: string
    type: object
...
info:
  contact:
    name: Containerlab API Server maintainers
    url: https://github.com/srl-labs/clab-api-server/issues
  description: |-
    REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.

    Most endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer <token>.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Containerlab API Server
  version: "1.0"
paths:
...
  /api/v1/labs/{labName}/nodes/{nodeName}/logs:
    get:
      description: |-
        Returns logs for a lab node.

        **Notes**
        - When `follow=true`, the response streams as NDJSON (one JSON object per line) until the client disconnects or the 30-minute timeout.
      parameters:
      - description: Name of the lab
        in: path
        name: labName
        required: true
        type: string
      - description: Full name of the container (node)
        in: path
        name: nodeName
        required: true
        type: string
      - default: all
        description: Number of lines to show from the end of logs (default all). Use
          an integer or 'all'.
        in: query
        name: tail
        type: string
      - description: 'Follow log output (stream logs as NDJSON). Note: In Swagger
          UI, streaming may not display correctly.'
        in: query
        name: follow
        type: boolean
      produces:
      - application/json
      - application/x-ndjson
      responses:
        "200":
          description: Container logs (follow=false). When follow=true, response is
            NDJSON stream of LogLine objects.
          schema:
            $ref: '#/definitions/models.LogsResponse'
        "400":
          description: Invalid input (lab name, node filter, etc.)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden (not owner of the lab)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Lab or node not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get node logs
      tags:
      - Logs
...
  /api/v1/labs/{labName}/topology/events:
    get:
      description: Streams topology YAML/annotations change events for a single lab
        topology document pair as NDJSON.
      parameters:
      - description: Lab name
        in: path
        name: labName
        required: true
        type: string
      - description: Relative topology YAML or annotations path inside lab directory
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: Topology document event stream
          schema:
            $ref: '#/definitions/models.TopologyDocEventResponse'
        "400":
          description: Invalid path
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stream topology document events
      tags:
      - Labs
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"